+++ examples/storage_system.tf
@@ -1,50 +1,53 @@
 provider "oneview" {
-        ov_username = "<ov-username>"
-        ov_password = "<ov-password>"
-        ov_endpoint = "<ov-endpoint>"
-        ov_sslverify = false
-        ov_apiversion = <ov-apiversion>
-        ov_ifmatch = "*"
+  ov_username   = "<ov_username>"
+  ov_password   = "<ov_password>"
+  ov_endpoint   = "<ov_endpoint>"
+  ov_sslverify  = false
+  ov_apiversion = <ov_apiversion>
+  ov_ifmatch    = "*"
 }
+
 
 resource "oneview_storage_system" "ss_inst" {
-        hostname = "<hostname>"
-        username = "<user>"
-        password = "<password>"
-        family   = "<family>"
+        hostname = "<storage_system_ip>"
+        username = "<storage_system_username>"
+        password = "<storage_system_password>"
+        family   = "StoreServ"
 }
 
 
-
-/*
 // Uncomment the following resource to update.
-resource "oneview_storage_system" "ss_inst"{
-        credentials = [
-        {
-		username = "<user>"
-	        password = "<password>"
-        }
-        ]
-        hostname = "<hostname>"
-        name = "<name>"
-        ports = [
-        {
-                mode = "<mode>"
-                id= "<id>"
-        }
-        ]
-        managed_pool = []
-        eTag = "--"
-        description = "TestStorageSystem"
-        uri = "/rest/storage-systems/<id>"
+/*resource "oneview_storage_system" "ss_inst" {
+  credentials = [
+    {
+      username = "<storage_system_username>"
+      password = "<storage_system_password>"
+    },
+  ]
+
+  hostname = "<storage_system_ip>"
+  name     = "ThreePAR-2"
+
+  storage_system_device_specific_attributes = {
+    managed_domain = "TestDomain"
+  }
+
+  eTag        = "--"
+  description = "TestStorageSystem"
+  uri         = "/rest/storage-systems/TXQ1010307"
 }
 */
-
-/* Testing the data source
-data "oneview_storage_system" "storage_system" {
-        name = "Cluster-2"
+/*
+// Testing the data source
+data "oneview_storage_system" "ss_int" {
+        name = "ThreePAR-2"
 }
 
 output "oneview_storage_system_value" {
-        value = "${data.oneview_storage_system.storage_system.uri}"
+        value = "${data.oneview_storage_system.ss_int.uri}"
+}
+// Testing import of existing resource
+*/
+/*resource "oneview_storage_system" "ss_import"{
 }*/
+