+++ CHANGELOG.md
@@ -11,7 +11,25 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Ethernet Network
 - FC Network
 - FCOE Network
+- Hypervisor Cluster Profile
+- Hypervisor Manager
+- Interconnect
+- Interconnect Type
+- Logical Enclosure
+- Logical Interconnect
+- Logical Interconnect Group
+- Network Set
 - Scopes
+- Server Certificates
+- Server Hardware
+- Server Hardware Type
+- Server Profile
+- Server Profile Template
+- Storage Pool
+- Storage System
+- Storage Volume
+- Storage Volume Attachment
+- Storage Volume Template
 - Uplink Set
 
 # [v1.3.0]